# openapi.yaml
openapi: 3.0.3
info:
  title: MintaiDialect API
  description: 闽台方言语音识别与数字嘉庚对话接口，目前对话可以使用普通话和闽南话作为输入，ASR和TTS暂时只支持闽南语音频和文本互转
  version: 1.0.0
servers:
  - url: http://115.190.116.189
security:
  - BearerAuth: ["admin-token-PbwEQUSPXLY"]
paths:
...
  /api/asr-tts/asr:
    post:
      summary: 语音识别 (ASR)
      description: 上传音频文件并指定语种，调用 ASR 服务进行语音识别，返回识别文本与真实音频时长（秒）。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ASRRequest"
      responses:
        "200":
          description: 识别成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponseASR"
              example:
                success: true
                message: "语音识别完成"
                data:
                  text: "你好，世界"
                  duration: 2.53
        "400":
          description: 参数或文件不合法
        "502":
          description: ASR 服务调用失败
  /api/asr-tts/tts:
    post:
      summary: 文本转语音 (TTS)
      description: 将文本转换为目标方言语音，返回对应的 POJ 发音、可播放的音频 URL与音频时长。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TTSRequest"
      responses:
        "200":
          description: 合成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponseTTS"
              example:
                success: true
                message: "文本转语音完成"
                data:
                  audio_url: "/uploads/tts_xxx.wav"
                  audio_duration: 1.68
                  file_size: 123456
                  audio_format: "wav"
                  poj_text: "Li2-ho2 se3-kai1"
        "400":
          description: 参数不合法（如文本长度超过1000字符）
        "502":
          description: LLM 或 TTS 服务调用失败
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # 可选，表示 token 格式
  schemas:
...
    ASRRequest:
      type: object
      required:
        - audio_file
        - source_language
      properties:
        audio_file:
          type: string
          format: binary
          description: 音频文件（支持 wav/mp3/flac/m4a/ogg 格式），文件大小限制10MB
        source_language:
          type: string
          enum: [minnan]
          default: minnan
          description: 语音识别语种（目前只支持minnan）
    BaseResponseASR:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "语音识别完成"
        data:
          type: object
          properties:
            text:
              type: string
              example: "你好，世界"
              description: 识别的文本
            duration:
              type: number
              format: float
              example: 2.53
              description: 真实音频时长（秒）
    TTSRequest:
      type: object
      required:
        - text
        - target_language
        - audio_format
      properties:
        text:
          type: string
          example: "你好世界"
          description: 要合成的文本（最大1000字符）
          maxLength: 1000
        target_language:
          type: string
          enum: [minnan]
          example: "minnan"
          description: 目标语言(目前只支持minnan)
        speed:
          type: number
          format: float
          example: 1.0
          description: 语速，默认 1.0
          minimum: 0.5
          maximum: 2.0
        audio_format:
          type: string
          enum: [wav, mp3, flac, m4a, ogg]
          example: "wav"
          description: 输出音频格式，如 wav, mp3, flac, m4a, ogg
    BaseResponseTTS:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功
          example: true
        message:
          type: string
          description: 消息
          example: "文本转语音完成"
        data:
          type: object
          properties:
            audio_url:
              description: 音频URL
              type: string
              example: "/uploads/tts_xxx.wav"
            audio_duration:
              description: 音频时长（估算）
              type: number
              format: float
              example: 1.68
            file_size:
              description: 文件大小（字节）
              type: integer
              example: 123456
            audio_format:
              description: 实际音频格式
              type: string
              example: "wav"
            poj_text:
              description: POJ发音
              type: string
              example: "Li2-ho2 se3-kai1"